[EH] modify : separate encryption & generateRandom as middleware
